fix(ci): python-coverage-shards — split two oversized test files under maintainability line limit
test_cleanup_branch_edges.py (1553) and
test_cleanup_auth_overlay_retry_part_002.py (1508) both exceeded the 1500-line
guardrail in test_first_party_code_files_stay_under_line_limit after recent
#583/#584 work, failing python-coverage-shards (8).

Move the self-contained _load_terminal_runtime_candidate guard-branch tests into
a new sibling test_cleanup_branch_edges_load_candidate.py, and move the
revoke-race / cross-cycle marker-write guards (sections 8 + 8b) into a new
test_cleanup_auth_overlay_retry_part_003.py that reuses the shared part 001/002
fixtures and helpers. No test behavior changed.
